requirements.txt
torch
torchvision
Pillow
streamlit
numpy
requests

